refactor: Split pucoti in multiple files
